Massive Redo
Changed the setup a great deal to accommodate larger maps and a
multiview system based off a single map. A lot of it is pretty abstract
at the moment, so let me know if you need any explanation. The redo is
far from done, so the Screen and one test are bombing. Seriously, if
anybody can find out why that one Map test is bombing, that would make
me a happy panda.