Stage creation/display system redone
Now stages can be created through a constructor receiving a String...
that represents the stage. The Map array now stores enums that are
procedurally drawn by the Screen ACCORDING TO THE LOCAL ORIGIN. This
will allow us to draw sections of the Map according to what section is
currently relevant. The origin variable will increase and decrease by
SCREENDIM when the player leaves the screen. The screen will then redraw
according to the current origin value. Enjoy!