Bugged
The last commit was bugged, so I have committed my proposed fix to the
bug, only to encounter another bug. The moveEnemies() method is causing
unnecessary death to the Player. The checkCurrentStatus method caused
normal Player death when called instead of the checkEnemyCollision()
method, minus double jumping. There is an image change in the
checkEnemyCollision() method simply for debugging purposes (to show when
the statement is executed) that can be removed after debugging. The
die() statement in said method always seems to be executed. Even when I
change the != null in the if statement to a == null! This is baffling
me. Help?